+++ src/attacks/attack_non_numeric.c
@@ -6,7 +6,7 @@
 #include "../header_fields.h" 
 #include "attack_non_numeric.h"
 
-bool attack_non_numeric(const char *output_filename, int index) {
+bool attack_non_numeric(const char *output_filename, uint8_t index) {
     bool is_header_tested = true;
 
     tar_archive archive;
